room

+++ back-end/liargame/src/main/java/com/sixsense/liargame/api/controller/SubjectController.java
@@ -23,7 +23,7 @@ public class SubjectController {
     }
 
     @GetMapping("/{subjectId}/words")
-    public ResponseEntity<?> getWord(@PathVariable Long subjectId, Integer roomId) {
+    public ResponseEntity<?> getWord(@PathVariable Long subjectId, Long roomId) {
         subjectService.selectRandomWord(subjectId, roomId);
         return ResponseEntity.ok().build();
     }
